requirements.txt
mysql-connector-python
pymongo
pytest 
pytest-cov 
unittest-mock
black 
ruff
python-dotenv